Change validation to use FIPS compliant settings
